Filter stocks that make it into the database.
We don't want every possible stock out there, we just want valid stocks with
clean data. Sometimes EOD or Polygon get some funds confused as stocks, or
delisted companies, etc.

Now the results of both are intersected and those are the ones that make it in,
this way we try to maximize stocks with valid consistent data.